cluster: use partition_manifest_downloader in topic recovery validation
Before allowing a recovery topic creation, the controller leader will
perform some validation on the partitions and segments.

During this process, Redpanda may not know the format of the partition
manifests (serde or json), and segment paths may include a cluster uuid
moving forward.

The former is handled transparently by the partition manifest
downloader, and the latter can be addressed by using a path provider
directly.