cloud_storage: use path provider + downloader in scrubber/anomaly detector
The anomaly detector (and by extension, the scubber)  needs to be able
to find a partition manifest whose format (serde or json) is not yet
known. This is handled transparently by the partition manifest
downlaoder, while also taking remote labels into account.

This commit updates anomaly detector to use the downloader, and plugs in
the archiver's path provider from the scrubber.