cloud_storage: simplify semantics of topic manifest serialization
Now that version is serialized via the implicit envelope version, we no
longer need to track version in memory. It has no semantic influence
over anything and just serves to complicate the interface.

This commit removes it, and removes the get_manifest_format_and_path()
method, with the expectation that the topic manifest _doesn't_ know what
format it should serialize, and it's the callers that should make this
determination.

Also note, previously the topic manifest serialization was tied to the
manifest object itself via the tracked version, but now that that's no
longer the case, so topic_manifest::serialize() is always going to be
expected to return binary format. topic_manifest::update() has been
changed to accommodate this.