cloud_storage: clean up manifest path generation from manifests
Topic/partition manifests are by far the biggest offender of things that
generate paths without the path provider that really should be using the
path provider (because their paths naturally collide across clusters in
the same bucket).

Thus far, we've been switching the important usages of path generation
to use the path provider, which helps avoid collisions. This commit
tightens things up a bit, cleaning up the remaining usages of
manifest-generated paths.

This means removal of:
- static manifest path generation methods
- format-focused interfaces, which were generally not using the format
  for anything important (mostly logging)
- methods in the manifest mentioning "legacy" formats (now
  subsumed by path utils and the path provider)
- importantly, remote::upload_manifest(), the most important offender of
  using paths generated by the manifest without consulting the path
  provider

Several tests and log lines are updated to accomodate this cleanup.