kafka: fix timequery failing for empty topics
This bug was introduced accidentally while trying to fix another
off-by-one bug in commit ref 8f2de964c0e915f4f10ae8eb74400e6288c5680f. I
forgot to account that for an empty topic the "start offset" is equal to
"last offset" so calling `model::prev_offset` would result in an offset
below the start which is invalid and throws an exception if passed
downstream.

To avoid the problem, we short-circuit with a "offset not found" response
straight away if that's the case.

https://github.com/redpanda-data/redpanda/pull/18112/commits/8f2de964c0e915f4f10ae8eb74400e6288c5680f