model/x: add kafka::offset_delta to offset_options::position variant
As a consequence of this, to maintain compatibility during partial
upgrades, introduce a legacy_transform_offset_options. This gives us
the ability to feature gate use of the new offset alternatives and avoid
writing non-forwards-compatible transform metadata in case a partial
upgrade needs to be rolled back.

This commit also adds transform_metadata::serde_write, which conditionally
writes the legacy version of the offset options struct iff the position
variant holds one of the legacy alternatives. This code can be removed after
v24.3 ships.

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>