c/shard_placement_table: better source shard state tracking in x-shard transfers
Previously if a cross-shard transfer failed, we couldn't really tell on
the source shard if we should retry or not (we may have failed to remove
obsolete state after a successful transfer, in this case retrying is
dangerous). Mark the state on the source shard obsolete immediately after a
successful transfer to fix that.

Also introduce more detailed failure conditions in prepare_tranfer() -
are we waiting for the source or the destination shard? This will come
handy when we implement moving data from extra shards because we'll have
to clean the destination ourselves.