rp_storage_tool: account for labels
Upcoming changes will enable supplying a cluster UUID to a topic to
change the remote paths that get used. To account for this, this changes
rp_storage_tool to look for the new path format and collect the cluster
UUID labels. This will allow us to continue running rp_storage_tool in
ducktape tests once the feature is enabled.