rptest: support remote labels in BucketView
BucketViews are typically used to examine the state of a bucket in the
context of a test. With objects proceeding to be include a given
cluster's UUID, this updates the view to know how to interpret the UUID.

This means most metadata scructures in the BucketView now account for a
label. I tried to not be too disruptive to existing tests, so for calls
that get various remote objects, we'll return the only one that exists.
If multiple exist (under multiple labels), an exception is thrown. This
is not expected to be the case until we begin sharing buckets in CI.