r/consensus: always use the latest configuration for followers metadata
When install snapshot request in processed by the follower it may not
always replace the follower log content. If the snapshot last included
offset is smaller than the follower dirty offset the follower should
prefix truncate all the data up to the snapshot last included offset but
keep all the entries which offset is greater than snapshot last included
offset.

Fixed the update of follower metadata state as it was always using the
snapshot configuration instead the latest from the configuration
manager.

Fixes: #core-internal/issues/1310

Signed-off-by: Michał Maślanka <michal@redpanda.com>