c/migrations: data migration worker service
Service to run on each shard of each node to perform operations on partitions.
Once data migration backend determines the shard where a partition replica
resides, it delegates the actual work to the worker instance on the shard.
It remains worker's responsibility to figure out whether it's a leader
of the partition raft and to react on partition leadership changes.