storage: fallocate at least one chunk to avoid appender hangs
In the `segment_appender::do_append` method we call
`do_next_adaptive_fallocation` repeatedly until there is enough space in
the file for the incoming operation. If `segment_fallocation_step` is
set to 0 then this would result in an infinite loop.

The code comments hint that there is another case when the fallocation
step is 0, when the disk is full. However,
`storage_resources::calc_falloc_step` clamps the minimum to append chunk
size too:

```
// At the minimum, falloc one chunk's worth of space.
if (step < min_falloc_step) {
    // If we have less than the minimum step, don't both falloc'ing at all.
    step = _append_chunk_size;
}
```