cloud_storage: add partition manifest paths to path provider
Adds partition manifest path generation to the remote path provider.

Note, in various places, we try to get the path of a partition manifest
but pass in a spillover manifest. This works today by having spillover
overriding the path generation method. This commit continues this, by
making the filename of the manifest overridable.