bazel: avoid leaking avro include path to dependents
Avro .cc files include "header.h" without the "avro/" path prefix. We
want to expose the parent of "avro" so that users of the library include
with the expected "avro/header.h" path, but don't want to expose the
non-prefixed path. The way to hide this would be to add `-I<path>` to
`copts`, but such a solution is ellusive (the recommended
`-Iexternal/...` doesn't seem to work as expected with bazel modules).

Anyway, the work around in this PR stages a copy of all the headers into
a separate directory and uses that as a set of private headers added to
`srcs`.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>