cloud_storage: clarify download_manifest naming
I have some upcoming changes that will handle binary/json ambiguity of
partition/topic manifest downloads outside of the remote. This will be
important, as such logic will also need to change as the naming scheme
is changed.

To that end, this changes the names of some manifest download methods to
be explicit about whether they're using binary or json format.