rpk: add additional disk self tests
Add 16K block size disk tests, a common block size written by Redpanda,
at varying IO depths: 1, 8 and 32 times the shard count (the
multiplication by the shard count happens in Redpanda and is
inevitable).

This will help better assess the performance of block storage which is
a bit outside the usual, in particular how it response to io depth
changes.

Additionally, add a 4K test which is the same as the existing one but
with dsync off. This is critical to assess the impact of fdatasync
on the storage layer: locally, this makes a 257x difference in
throughput though the effect is much more muted, perhaps close to zero
on other SSD types.

Rename slightly the tests to remove extraneous info.

Issue redpanda-data/core-internal#1266.