tests: do not set transaction timeout in tests that do not need it
Most of our transactions tests were build based with the same code which
often were copy pasted. The test were setting the default transaction
timeout to 10 seconds as they might have been based on the test that
validated the transactions expiration. Updated the tests to make sure
that the timeout is set in the test that need it.

The timeout made some tests to fail intermittently.

Fixes: #18184

Signed-off-by: Michał Maślanka <michal@redpanda.com>