cloud_storage: topic_manifest_downloader::find_manifests
This adds a find_manifests() method to the topic_manifest_downloader.
This will subsume the logic currently in topic_recovery_service, that
lists objects in the buckets looking for objects that look like topic
manifests. With our object naming scheme changing, this will need to
account for the new naming scheme.

Admittedly, there's somewhat of an impedence mismatch since the
topic_manifest_downloader expects to target one topic, so this just
makes the method static. But it still feels natural for this logic and
its tests to be coupled to topic_manifest_downloader.

The code isn't as optimal as it could be in terms of number of list/get
operations performed, since I wanted to rely on the downloader to
duplicating some of the path-related complexity. But this endpoint is
likely not going to be widely used, particularly as we begin having more
clusters per bucket. Instead controller snapshot-based restore will be
used.