cloud_storage: get topic_namespace from manifest paths
Adds a couple utility methods to get the topic namespace from a given
path. This will be useful in an upcoming commit where we do this to
parse list results.

Why not have a single method? We could, but the upcoming use case is
topic recovery, which uses a specific path prefix, and will know whether
it expects labeled or hash prefixed objects. In that context, it seems
natural to expect the correct kind of path.