bazel: allow rules_python to run with root user
This isn't ideal, but it sounds like it really only affects the ability
to cache *.pyc files. In our case, we have about a grand total of 3 or 4
files we compile with Python so I'm not sure it's worth the headache
right of figuring out a different way to deal with this.

https://github.com/bazelbuild/rules_python/issues/1169

Useful for docker builds that require extra effort to use in a non-root
context.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>