archival_stm: add remote path provider to archival stm
This plumbs the remote label (cluster uuid) associated with a topic from
the topics table into the archival meta stm, in the same fashion that
vcluster-id is plumbed into the rm_stm.

Conceptually, the archival STM will be the owner of the root remote path
provider used throughout the cloud_storage and archival subsystems.