model/x: Introduce transform_from_{start,end}
Thin wrappers around a kafka::offset_delta meant for use in a sum
type to indicate how the delta should be applied - i.e. added to
the start offset of a topic or subtracted from the latest offset
of a topic.

Enclosed delta should always be greater than or equal to zero.

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>