bazel: build xxhash explicitly
The version of xxhash provided by bzlmod requires -iquoted includes
which prevents redpanda's use of brackets: <xxhash.h>. this isn't
inherently a problem, we probably should use quoted includes. but in the
interim state where both cmake and bazel are in use we need some
flexibility. so we build xxhash manually so we can control this.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>