tests: added retries to raft force reconfiguration test
Raft force reconfiguration test is designed to skip one of the nodes
when force updating configuration. In this scenario it may sometimes
happen that the node which wasn't updated with the configuration
override will become a leader. In this situation controller backend in
Redpanda will retry. Added retries to the test to make sure the test
will retry reconfiguration if it was reverted by node which was missing
update.

Fixes: #19938

Signed-off-by: Michał Maślanka <michal@redpanda.com>
(cherry picked from commit 30efad6b3aa5258611b9dbbb5d0425d4eaa0aba2)