archival: Update archiver scheduler
Pass next housekeeping and next manifest upload to the scheduler. This
is needed to avoid sleeping too much when beckoff interval is large but
we need to do housekeeping soon or we need to upload the manifest. The
scheduler computes actual sleep interval based on the current backoff
interval and both timestamps.

Also, this commits makes any call use at least one unit from put
requests token bucket. This is needed to block in a busy-loop situation.