cluster: cache kafka offset of last applied record in log_eviction_stm
This allows the log_eviction_stm to know what the current kafka start
offset override is even if it doesn't reside in the local log. However,
since this cached value is not snapshotted by the stm it may not be
recovered when repanda restarts. In this case one would need to fallback
on the archival stm to find the kafka start offset override.

(cherry picked from commit feeec639f2d52379d33100c505682113deb03817)