tests: made atomic consume-transform-produce test more robust
The exactly once delivery semantic tests was using a stop condition
where a size of the finished processing partition set was compared with
the size of high watermark list. Sometimes the high watermark list may
not contain all the partitions as they are not reported when there is no
leader. This lead to test stopping to early and desitination topic
missing some messages.  Changed the test to use the topic partition
count in the stop condition instead.

Fixes: #20315

Signed-off-by: Michał Maślanka <michal@redpanda.com>