bazel/rc: refined release configuration
Default bazel copts for the `opt` `compilation_mode` set definition for
`FORTIFY_SOURCE` flag to `1`. In Redpanda we want the flag to be set to
`2`. Even when we undefine the flag it is still added twice to the
command line as it is set with the `compilation_mode`. This leads to
redefinition. Added our custom set of flags for optimized build.

Note:
The set of flags can be removed when we go with the `toolchain_llvm` and
we may add then to `opt_compile_flags` toolchain module parameter.

Signed-off-by: Michał Maślanka <michal@redpanda.com>