group/tx/tests: deflake group_tx_compaction_test
When a group boots up and recovers, a version fence batch is
asynchronously written to the log. In slow debug builds, this can happen
in the middle of a test run and is conflicting with offset expecations.
This fix waits for the version fence batch to be written before the
test execution begins.