r/hb_manager: use async_for_each when collecting heartbeat requests
When collecting heartbeat requests the heartbeat manager must iterate
over every single raft instance on the shard and then over all followers
of instances which are the leader. The request collection logic is now
synchronous. For now the loop do not cause reactor stalls but as we
increase the partition density it may eventually lead to stalling
reactor.

Changed the implementation to use `ssx::async_for_each` to yield
occasionally if needed.

Signed-off-by: Michał Maślanka <michal@redpanda.com>