pandaproxy/schema_registry: is_superset break infinite recursion
eventually, with default values for properties being "true/false" or
"{}/{"not":{}}", we can form a call like
is_superset("{}", ...) or is_superset(..., "{"not":{}") that can cause
infinite recursion. this is because the type is assumed to mean every
type, and this can lead to this recursive chain is_superset ->
is_array_superset -> is_superset ...

this commit breaks the chain by recognizing these atoms and returning
early.

normalization and/or AST could offer a more robust solution, by ensuring
that implicit properties are explicit and removing extraneous fields.

The test cases in this commit would lead to infinite recursion without
the early return in is_superset.