schema_registry/json: split validate_json_schema from parse_json
this function gets a dialect and a json::Document of the candidate
schema, and will perform validation against the correct metaschema.

to compile only once the Schema object,
get_metaschema<json_schema_dialect>() will compile it lazily and save it in static const
variable.

to support validation of json without the "$schema" property,
try_validate_json_schema will try to validate it against the known
dialects, from the most recent to the oldest