tx/tests: deflake test_mark_transaction_expired
The test does duplicate expirations which results in a ERROR log
in the brokers. The duplicate expirations stem from the fact that the
previous test logic explicitly marked a producer for expiration on every
ntp involved in the transaction. This is not needed as the expiration is
routed through the coordinator and that aborts the transaction on all
participants. Refactored the test logic accordingly.