cloud tests: run all OMB validation tests
This test was disabled while we were trying to get all the cloud
tests green and we had the latency spike issue. Per
redpanda-data/redpanda#15334 we now have retry logic and in any case
this test is only "just as affected" as test_common which are
successfully running (for the most part).

So re-enable it by specifying the entire OMBValidationTest class in the
suite.