cloud_storage: log more context from topic_manifest_downloader
The downloader has a lot of context but reduces results into an outcome
or error enum. This updates the downloader to log before returning early
from the class, logging errors at ERROR log level, and anything else at
INFO level (since it should be up to callers whether such code paths are
actually problematic).

This is review follow-up to
https://github.com/redpanda-data/redpanda/pull/20778