cst/inv: Add inventory scrub config properties
A set of properties are added to support and enable inventory based
scrub. This feature is off by default.

A property called cloud_storage_inventory_self_managed_report_config is
added to facilitate testing and inventory based scrub on backends which
do not support automatic reports. If this is enabled redpanda will not
attempt to schedule reports on startup. We will however still
periodically check if a report is available in the expected path using
the same code path that is used for S3. If a report is found there it
will be consumed.