cluster: add `topic_namespace_override` to `topic_properties`
To allow for an aliased `topic_namespace` for use with tiered storage,
that differs from the `ntp()` locally configured for a topic.

This `topic_namespace_override` will be passed to the `remote_path_provider`
constructor in future commits.

Also provides a helper function `archival_tp_ns()`, which returns the
appropriate `topic_namespace` to users of the `topic_configuration` for
archival purposes.