kafka: fix timequery when local log is empty
Before querying local log we bump the `timequery_config::min_offset`.
This offset is inclusive. When local log is empty, the min_offset points
one unit above the end of the log. In this PR, we add a check to return
early when this is the case.

(cherry picked from commit 3dd1c3a45432c6019b1e8c154dc47751fefd035b)