k/alter_configs: fix topic config resetting
The AlterConfigs API differs from the IncrementalAlterConfigs API in
that it is supposed to clear all configs not specified in the request.
This was implemented incorrectly, so this commit fixes the behaviour by:
 * `op_t::delete`'ing unspecified fields instead of `op_t::set`'ing them
   to their type's default value.
 * Doing this for all topic configs, including the currently missing
   ones:
    - `remote_delete`
    - `segment_ms`
    - `batch_max_bytes`
    - `retention_local_target_ms`
    - `retention_local_target_bytes`
    - `initial_retention_local_target_ms`
    - `initial_retention_local_target_bytes`
    - `write_caching`
    - `flush_ms`
    - `flush_bytes`
 * Fixing up the shadow indexing mode parsing method because this
   assumed that `properties.shadow_indexing_mode.op` was always `set` by
   default which is no longer the case.

Note that an exception is made for the shadow indexing topic configs
`redpanda.remote.(read|write|delete)` to not reset them by this API.
This is done in the next commit.