archival: bump offset if seeking undershoots
convert_begin_offset_to_file_pos can undershoot, in which case the
metadata returned by the archival policy won't perfectly align with the
caller's expectations. This could result in an offset_overlap when
applying the upload metadata.

It's unclear exactly why there are missing batches, given this code is
used for the non-compacted uploader. But the fix is simple enough, and
is one that we use in the compacted upload policy already[1].

[1] https://github.com/redpanda-data/redpanda/blob/4b8135ea5dbdca322ccf8efc85424413a20c5ade/src/v/archival/segment_reupload.cc#L516-L522

(cherry picked from commit 8b616df65bfbddb5dc526bb1e9313ee8d125e81f)