crypto/bench: Only load context once per run
OpenSSL doesn't appear to like being initialized, uninitialized, and
then re-initialized by the same process over and over again.  This
results in some memory being left around and reused even after being
freed which is what was causing the crypto bench test to segfault and
fail.  The solution is to only bring up the OpenSSL context service once
per run and then bring it down when the Seastar engine is finishing up.

Fixes: CORE-4207

Signed-off-by: Michael Boquard <michael@redpanda.com>