tests: increase tolerance in local retention movement test
Increased tolerance for the size of partitions on a node added to the
cluster. The expected partition size may vary as it is dependent from
the segment size and non log data. The tolerance allow us to account for
the variability without compromising the test correctness.

Fixes: #20225

Signed-off-by: Michał Maślanka <michal@redpanda.com>
(cherry picked from commit cc099592dcc707e19582b30279e6334342ad8081)