c/shard_placement_table: fix the notification order in set_target()
Previously, when requesting an x-shard transfer, we first notified the
source shard, and then the destination. This led to an unnecessary wait,
as the first reconciliation attempt on the source shard was unsuccessful
(destination was not yet ready). Change the order so that the transfer
can proceed immediately in the most common case.

This is not required for correctness, only for performance.