cluster: pass xshard transfer state from old partition instance to new
To make starting a partition instance on the destination shard faster,
pass a bit of in-memory state from the old partition. This is strictly
an optimization, but can make e.g. regaining leadership faster after the
transfer. Another possible use: pass in-memory state of STMs or offset
translator to avoid re-reading it from disk.

This commit contains only the required plumbing, actual substance to
this struct will be added in later commits.