c/topic: use batch cache for transaction coordinator topic
The transactional coordinator state machine is waiting for application
of batches to the state when handling client requests it is critical for
the application to happen timely. Previously the coordinator topic
wasn't using `storage::batch_cache` which made the appended batches not
immediately available to read by the state machine manager. In this case
the manager waits for some time to prevent busy looping.

Enabling the cache for the transaction manager topic makes the
operations much faster and decreases the client latency.

Signed-off-by: Michał Maślanka <michal@redpanda.com>