schema_registry/json: is_positive_combinator_superset base
this check is meant to cover anyOf/oneOf/allOf.

base function that just check that no combinator is used by both schemas

note: json schema allows more than one combinator to appear in a schema,
but this it's difficult to define how is_superset should behave in this
case. since this seems to be an edge case, this implementation will
simply throw an exception.
the schemas themself can still be used but it's not possible to perform
compatibility checks on them