schema_registry/json: is_positive_combinator_superset stub mismatch
given the current implementation, a combinator can have 4 states
not set, allOf, oneOf, anyOf.

in case of state mismatch between schema, this commit throws an
exception as not implemented (that gets converted to false)

this is overly restrictive:
some combinations, like (not set, anyOf), or (anyOf, not set) or cases
where the schema array has only one value, might be compatible.

this will be revisited in a later commit