bazel: disable github workflow for pull requests
There is an issue loading secrets from repository forks. The only real viable solution is to ask all engineers to submit PRs from branches pushed to the upstream repo, but this is a major change given that we'll move to buildkite soon anyway. So the plan now is to just re-home the workflow to buildkite in the short term.