schema_registry/json: is_positive_combinator_superset subschemas matching
once the function has two schema arrays of suitable cardinality, we need
to check that every schema of the smaler set has a distinct match in the
other set.

this is solved in a general way by constructing a bipartite graph, and
computing the maximum cardinality matching on the graphs.

an example:
older= {"anyOf": [A, B, C, D]}
newer= {"anyOf": [1, 2, 3]}

for each pair
o ∈ [A, B, C, D]
n ∈ [1, 2, 3]
if (is_superset(o, n)) then we add a edge between o and n

so we find out that the list of edges is
[(A, 2), (A, 3), (B,1), (C, 2)],

one maximum cardinality matching is
[(A, 3), (B, 1), (C, 2)].
every subschema from newer (them smaller set) has a distinct companion
in older, meaning that we are sure that is_superset(older, newer) every
json matched by one of the subschemas of newer will be matched by a
subschema of older