schema_registry/json: relax multipleOf check
increase the threshold to accept
{"type": "number", "multipleOf": 1.1}
{"type": "number", "multipleOf": 3.3}

std::abs(std::reminder(3.3, 1.1)) is ~5e-16

this is a fairly small value, but extremely bigger than the previous
10*epsilon limit

the new threshold depends on the Unit of Least Precision (ULP) relative
to the bigger value.
This value is the distance to the next floating
point value, and RapidJson documents that it's string->double parsing
code has at most a 3*ulp error. For this reason, this is the threshold
used to check if the reminder is close enough to zero

(cherry picked from commit 14c82a5741bbb9ef1cf25b3fbeef14e344f1b082)