rptest/kgo: avoid extra wait on shutdown
There are 2 ways to stop a kgo service. One is graceful shutdown and
another one is hard stop. We use 2 `threading.Event` objects to signal
this.

Before this change, the polling thread would only wait on the
`_stop_requested` and if it was set or if a timeout is hit
(self.INTERVAL=5s) it would run the loop again and check if stop or
shutdown was requested.

This leads for shutdown taking 5s more than it should in worst case.

We avoid this extra wait but settling on a single threading.Event object
(_shutdown_requested) for signaling that the loop should run again. Then
we signal that event from both, shutdown and hard stop.