storage: plumb abort source to compaction copy reducer
We've seen that during compaction, the step that reads a segment and
deduplicates each record can take a long time. During this span, we do
not check the compaction abort source, and this can result in
compactions blocking partition shutdown.

This commit plumbs the compaction abort source into the copy reducer and
uses it for both self compaction and windowed compaction.