dt/x: Clean up log manager metrics values test
Previously, the test would produce some random data to the input topic,
leading to an arbitrary distribution of records across partitions and
an upredictable distribution of buffered log data across nodes. The test
condition was whether any log buffer usage increased as a result but
relied on some obscure behavior of the franz-go consumer to verify that
the produce was making progress.

The test has been reformulated to instead produce enough data to exceed
log buffer capacity on a given partition, and the assertion softened to
ensure that no metrics indicate buffers overflowing.

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>